스크롤 네비게이션으로 변경
페이지 전환 방식을 해쉬라우팅에서 스크롤 네비게이션으로 변경
