+++ client/package.json
@@ -11,6 +11,7 @@
     "react-mobx": "^0.0.4",
     "react-router-dom": "^5.0.0",
     "react-scripts": "2.1.1",
+    "react-scroll": "^1.7.11",
     "reactstrap": "^7.1.0"
   },
   "scripts": {
