+++ client/src/components/Template.js
@@ -1,33 +1,33 @@
-import React, { Component } from "react";
-import { Container, Row, Col } from "reactstrap";
-import Header from './header/Header'
-import Footer from './footer/Footer'
+// import React, { Component } from "react";
+// import { Container, Row, Col } from "reactstrap";
+// import Header from './header/Header'
+// import Footer from './footer/Footer'
 
-export default class Template extends Component {
+// export default class Template extends Component {
     
-  render(){
+//   render(){
 
-    const {children, projects} = this.props
+//     const {children, projects} = this.props
     
-    return(
-      <Container fluid>
-        {/* header */}
-        <Row className="header">
-          <Col>
-            <Header projects={projects}/>
-          </Col>
-        </Row>
-          {/* body */}
-        <Row className="body">
-          <Col className="bodyArea">{children}</Col>
-        </Row>
-        {/* footer */}
-        <Row className="footer">
-          <Col>
-            <Footer/>
-          </Col>
-        </Row>
-      </Container>
-    );
-  }
-}
+//     return(
+//       <Container fluid>
+//         {/* header */}
+//         <Row className="header">
+//           <Col>
+//             <Header projects={projects}/>
+//           </Col>
+//         </Row>
+//           {/* body */}
+//         <Row className="body">
+//           <Col className="bodyArea">{children}</Col>
+//         </Row>
+//         {/* footer */}
+//         <Row className="footer">
+//           <Col>
+            
+//           </Col>
+//         </Row>
+//       </Container>
+//     );
+//   }
+// }
