+++ client/src/components/body/elements/Profile.js
@@ -1,0 +1,32 @@
+import React, { Component } from 'react';
+import Template from "../../Template";
+import { Container, Row, Col, 
+        ListGroup, ListGroupItem } from "reactstrap";
+import "../../../static/css/profile.css";
+import photo from "../../../static/images/logo.svg"
+import {Element} from 'react-scroll';
+
+export default class Profile extends Component{
+    render(){
+        return(
+            <Element name="profile">
+                <div>
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                </div>
+            </Element>
+            
+        )
+    }
+}
