+++ client/src/components/body/elements/Projects.js
@@ -1,0 +1,27 @@
+import React, { Component } from 'react';
+import Template from "../../Template";
+import {Element} from 'react-scroll';
+
+export default class Projects extends Component{
+    render(){
+        return(
+            <Element name="projects">
+                <div>
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                    element1
+                </div>
+            </Element>
+        )
+    }
+}
