--- client/src/components/body/pages/Projects.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-import Template from "../../Template";
-
-export default class Projects extends Component{
-    render(){
-        return(
-            <Template>
-                <div>this is Projects page</div>
-            </Template>
-        )
-    }
-}
