--- client/src/components/body/pages/Skills.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-import Template from "../../Template";
-
-export default class Skills extends Component{
-    render(){
-        return(
-            <Template>
-                <div>this is skills page</div>
-            </Template>
-        )
-    }
-}
