+++ client/src/components/header/Header.js
@@ -1,5 +1,6 @@
 import React, { Component } from "react";
-import { Link } from "react-router-dom";
+// import { Link } from "react-router-dom";
+import { Link } from "react-scroll";
 import { Nav, NavItem, NavLink, 
         Dropdown, DropdownItem, DropdownToggle, DropdownMenu,} from "reactstrap";
 import '../../static/css/header.css';
@@ -29,23 +30,23 @@ export default class Header extends Component{
                 <div className="header-div header-nav">
                     <Nav>
                         <NavItem>
-                            <Link className="nav-link" to="/">main</Link>
+                            <Link className="nav-link" to="/" spy={true} smooth={true}>HOME</Link>
                         </NavItem>
                         <NavItem>
-                            <Link className="nav-link" to="/profile">profile</Link>
-                        </NavItem>
-                        <NavItem>
-                            <Link className="nav-link" to="/skills">skills</Link>
+                            <Link className="nav-link" to="/profile">ABOUT</Link>
                         </NavItem>
                         <Dropdown nav isOpen={this.state.dropdownOpen} toggle={this.toggle}>
                             <DropdownToggle nav caret>
-                                projects
+                                PROJECTS
                             </DropdownToggle>
                             <DropdownMenu>
                                 <DropdownItem>스터디타임</DropdownItem>
                                 <DropdownItem>인공위성 관제 모니터링 웹 서비스</DropdownItem>
                             </DropdownMenu>
                         </Dropdown>
+                        <NavItem>
+                            <Link className="nav-link" to="/skills">CONTACT</Link>
+                        </NavItem>
                     </Nav>
                 </div>
                 <hr/>
@@ -54,4 +55,39 @@ export default class Header extends Component{
             
         );
     }
+
+    // render(){
+    //     return(
+    //         <div>
+    //             <div className="header-div header-title">
+    //                 <span className="nav-link">Yoonho Shin</span>
+    //             </div>
+    //             <div className="header-div header-nav">
+    //                 <Nav>
+    //                     <NavItem>
+    //                         <Link className="nav-link" to="/">HOME</Link>
+    //                     </NavItem>
+    //                     <NavItem>
+    //                         <Link className="nav-link" to="/profile">ABOUT</Link>
+    //                     </NavItem>
+    //                     <Dropdown nav isOpen={this.state.dropdownOpen} toggle={this.toggle}>
+    //                         <DropdownToggle nav caret>
+    //                             PROJECTS
+    //                         </DropdownToggle>
+    //                         <DropdownMenu>
+    //                             <DropdownItem>스터디타임</DropdownItem>
+    //                             <DropdownItem>인공위성 관제 모니터링 웹 서비스</DropdownItem>
+    //                         </DropdownMenu>
+    //                     </Dropdown>
+    //                     <NavItem>
+    //                         <Link className="nav-link" to="/skills">CONTACT</Link>
+    //                     </NavItem>
+    //                 </Nav>
+    //             </div>
+    //             <hr/>
+    //         </div>
+
+            
+    //     );
+    // }
 }
