+++ client/src/index.js
@@ -9,9 +9,8 @@ import * as serviceWorker from './serviceWorker';
 import { HashRouter,Route } from "react-router-dom";
 
 import Main from "./components/body/pages/Main";
-import Profile from "./components/body/pages/Profile";
-import Projects from "./components/body/pages/Projects";
-import Skills from "./components/body/pages/Skills";
+// import Profile from "./components/body/pages/Profile";
+// import Projects from "./components/body/pages/Projects";
 
 window.React = React;
 
@@ -19,9 +18,8 @@ ReactDOM.render(
     <HashRouter>
         <div calssName="index">
             <Route exact path = "/" component={Main}/>
-            <Route path = "/profile" component={Profile}/>
-            <Route exact path = "/projects" component={Projects}/>
-            <Route exact path = "/Skills" component={Skills}/>
+            {/* <Route path = "/profile" component={Profile}/>
+            <Route path = "/projects" component={Projects}/> */}
         </div>
     </HashRouter>,document.getElementById('root')
 )
